refactor: reorganize project using FSD structure
