--- app/page.tsx
@@ -1,5 +1,0 @@
-import InteractiveSquareEditor from "../interactive-square-editor"
-
-export default function Page() {
-  return <InteractiveSquareEditor />
-}
