+++ src/app/page.tsx
@@ -1,0 +1,5 @@
+import { InteractiveSquareEditor } from "@/features/interactive-square-editor"
+
+export default function Page() {
+  return <InteractiveSquareEditor />
+}
