+++ src/entities/green-polygon/index.ts
@@ -1,0 +1,1 @@
+export { GreenPolygon } from './ui'
