+++ src/entities/inner-strip/index.ts
@@ -1,0 +1,1 @@
+export { InnerStrip } from './ui'
