+++ src/entities/inner-strip/ui.tsx
