+++ src/entities/outer-strip/index.ts
@@ -1,0 +1,1 @@
+export { OuterStrip } from './ui'
