+++ src/entities/outer-strip/ui.tsx
