+++ src/entities/road-strip/index.ts
@@ -1,0 +1,1 @@
+export { RoadStrip } from './ui'
