+++ src/features/interactive-square-editor/index.ts
@@ -1,0 +1,1 @@
+export { default as InteractiveSquareEditor } from './ui'
