+++ src/shared/ui/svg-shader-material.tsx
